+++ client/common/src/main/java/su/plo/voice/client/audio/capture/PushToTalkActivation.java
@@ -22,7 +22,7 @@ public final class PushToTalkActivation extends BaseActivation {
     }
 
     @Override
-    public @NotNull Result process(byte[] samples) {
+    public @NotNull Result process(short[] samples) {
         if (isDisabled() || !voiceClient.getCurrentServerInfo().isPresent())
             return Result.NOT_ACTIVATED;
